Rename QueryIdCachingProxyHandler to HttpUtils
