+++ gateway-ha/src/main/java/io/trino/gateway/ha/handler/HttpUtils.java
@@ -13,7 +13,7 @@
  */
 package io.trino.gateway.ha.handler;
 
-public class QueryIdCachingProxyHandler
+public class HttpUtils
 {
     public static final String V1_STATEMENT_PATH = "/v1/statement";
     public static final String V1_QUERY_PATH = "/v1/query";
@@ -26,5 +26,5 @@ public class QueryIdCachingProxyHandler
     public static final String OAUTH_PATH = "/oauth2";
     public static final String USER_HEADER = "X-Trino-User";
 
-    private QueryIdCachingProxyHandler() {}
+    private HttpUtils() {}
 }
